extract CropImageContractOptions into its own file
